fix: 从 gin ctx.get 获取的 userid 类型改为 int32

+++ service/user/login.go
@@ -21,7 +21,7 @@ func generateToken(user user.User) (token string, err error) {
 	//生成token
 	j := middlewares.NewJWT()
 	claims := models.CustomClaims{
-		ID:          uint(user.ID),
+		ID:          user.ID,
 		NickName:    user.NickName,
 		AuthorityId: uint(user.Role),
 		RegisteredClaims: jwt.RegisteredClaims{
